+++ StortSpelProjekt/Engine/src/AudioEngine/AudioVoice.cpp
@@ -2,6 +2,7 @@
 #include "AudioVoice.h"
 #include "AudioBuffer.h"
 #include "../Misc/Option.h"
+#include "../Events/EventBus.h"
 
 AudioVoice::AudioVoice()
 {
@@ -33,6 +34,8 @@ void AudioVoice::initialize(XAUDIO2_BUFFER* buff, WAVEFORMATEXTENSIBLE* wfxForma
 
     m_pSourceVoice->SetVolume(std::atof(Option::GetInstance().GetVariable("f_volume").c_str()));
     HRESULT hRes = m_pSourceVoice->SetOutputVoices(NULL);
+
+    Stop();
 }
 
 void AudioVoice::Initialize(AudioBuffer& audio)
@@ -77,5 +80,39 @@ void AudioVoice::Stop()
         // reset the buffer so the sound starts from the beginning at next playback
         m_pSourceVoice->FlushSourceBuffers();
         m_pSourceVoice->SubmitSourceBuffer(m_pBuffer);
+    }
+}
+
+void AudioVoice::Resume()
+{
+    if (m_Initialized)
+    {
+        // reset the buffer if sound has ended
+        XAUDIO2_VOICE_STATE test;
+        m_pSourceVoice->GetState(&test);
+        if (test.BuffersQueued == 0)
+        {
+            // reset the buffer so the sound starts from the beginning at next playback
+            m_pSourceVoice->FlushSourceBuffers();
+            m_pSourceVoice->SubmitSourceBuffer(m_pBuffer);
+        }
+
+        // play the sound
+        if (FAILED(m_pSourceVoice->Start(0)))
+        {
+            Log::Print("Error playing audio\n");
+        }
+    }
+}
+
+void AudioVoice::Pause()
+{
+    if (m_Initialized)
+    {
+        // stop playback
+        if (FAILED(m_pSourceVoice->Stop(0)))
+        {
+            Log::Print("Error stopping audio\n");
+        }
     }
 }
