+++ StortSpelProjekt/Engine/src/ECS/Components/Audio2DVoiceComponent.h
@@ -5,6 +5,7 @@
 #include <Windows.h>
 
 class AudioVoice;
+struct PauseGame;
 
 // Component used for playing 2D/background sounds
 namespace component
@@ -29,6 +30,9 @@ namespace component
 
 	private:
 		std::map<std::wstring, AudioVoice> m_Voices;
+		std::map<std::wstring, bool> m_WasPlaying;
+
+		void pauseToggleAudio(PauseGame* evnt);
 	};
 }
