+++ StortSpelProjekt/Engine/src/ECS/Scene.cpp
@@ -113,10 +113,6 @@ NavMesh* Scene::CreateNavMesh(const std::string& type)
     {
         m_pNavMesh = new NavMesh(this, type);
     }
-    else
-    {
-        Log::Print("Scene already has NavMesh\n");
-    }
     return m_pNavMesh;
 }
