+++ StortSpelProjekt/Engine/src/Renderer/DX12Tasks/CopyPerFrameTask.cpp
@@ -41,6 +41,7 @@ void CopyPerFrameTask::Execute()
 
 	m_pCommandInterface->Reset(m_CommandInterfaceIndex);
 
+	volatile int counter = 0;
 	for (auto& tuple : m_UploadDefaultData)
 	{
 		if (std::get<2>(tuple) != nullptr)
@@ -51,6 +52,11 @@ void CopyPerFrameTask::Execute()
 				std::get<1>(tuple),		// DefaultHeap
 				std::get<2>(tuple));	// Data
 		}
+		else
+		{
+			Log::Print("Tried to copy nullptr\n");
+		}
+		counter++;
 	}
 
 	commandList->Close();
