+++ StortSpelProjekt/Game/src/main.cpp
@@ -14,7 +14,7 @@
 #include "Components/UpgradeComponents/UpgradeComponent.h"
 #include "MainMenuHandler.h"
 #include "GameOverHandler.h"
-#include "UpgradeGUI.h"
+#include "PauseGUI.h"
 
 #include "Misc/Edge.h"
 
@@ -33,10 +33,6 @@ void ShopUpdateScene(SceneManager* sm, double dt);
 int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE, LPSTR, int nCmdShow)
 {
     _CrtSetDbgFlag(_CRTDBG_ALLOC_MEM_DF | _CRTDBG_LEAK_CHECK_DF);
-    //Cryptor::DecryptDirectory(Cryptor::GetGlobalKey(), "../Vendor/Resources/Models/CastleFloor");
-    //Cryptor::DecryptDirectory(Cryptor::GetGlobalKey(), "../Vendor/Resources/Models/StoneWallParts/Wall");
-    //Cryptor::EncryptDirectory(Cryptor::GetGlobalKey(), "../Vendor/Resources/Models/CastleFloor");
-    //Cryptor::EncryptDirectory(Cryptor::GetGlobalKey(), "../Vendor/Resources/Models/StoneWallParts/Wall");
     /*------ Load Option Variables ------*/
     Option* option = &Option::GetInstance();
     option->ReadFile();
@@ -72,7 +68,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE, LPSTR, int nCmdShow)
     GameNetwork gameNetwork;
 
     /*-------- UpgradeGUI ---------*/
-    UpgradeGUI::GetInstance().Init();
+    PauseGUI::GetInstance().Init();
 
     /*------ Network Init -----*/
 
@@ -92,7 +88,6 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE, LPSTR, int nCmdShow)
     {
         /* ------ Update ------ */
         timer->Update();
-        logicTimer += timer->GetDeltaTime();
 
         bool changedScene = sceneManager->ChangeScene();
         if(changedScene)
@@ -110,6 +105,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE, LPSTR, int nCmdShow)
 
         if (!Input::GetInstance().IsPaused())
         {
+            logicTimer += timer->GetDeltaTime();
             sceneManager->RenderUpdate(timer->GetDeltaTime());
             particleSystem->Update(timer->GetDeltaTime());
             if (logicTimer >= updateRate)
@@ -126,13 +122,12 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE, LPSTR, int nCmdShow)
                 sceneManager->Update(updateRate);
                 physics->Update(updateRate);
                 GameGUI::GetInstance().Update(updateRate, sceneManager->GetActiveScene());
-                UpgradeGUI::GetInstance().Update(updateRate, sceneManager->GetActiveScene());
+                PauseGUI::GetInstance().Update(updateRate, sceneManager->GetActiveScene());
             }
         }
         else
         {
-            GameGUI::GetInstance().Update(timer->GetDeltaTime(), sceneManager->GetActiveScene());
-            UpgradeGUI::GetInstance().Update(timer->GetDeltaTime(), sceneManager->GetActiveScene());
+            PauseGUI::GetInstance().Update(timer->GetDeltaTime(), sceneManager->GetActiveScene());
 
             /* ------ ImGui ------*/
             if (DEVELOPERMODE_DEVINTERFACE == true)
@@ -168,8 +163,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE, LPSTR, int nCmdShow)
 Scene* LoadScene(SceneManager* sm)
 {
     Scene* scene = sm->CreateScene("LoadScene");
-    std::vector<float3> spawnPoints;
-    AssetLoader::Get()->GenerateMap(scene, "../Vendor/Resources/Rooms", &spawnPoints, { 3.0f,3.0f }, { 173.0f,200.0f }, false);
+    AssetLoader::Get()->LoadAllMaps(scene, "../Vendor/Resources/Rooms");
     return scene;
 }
